Update WinPreManage_browsers_history.ps1
Commit Message:
----------------
Enhanced Firefox History Backup and User Interaction in Browser Backup Utility

- Modified the `BackupFirefoxHistory` method to dynamically identify and backup history from Firefox profiles located in the user's AppData\Roaming directory. This enhancement addresses the issue of not finding the Firefox history database due to variations in profile paths across different installations.

- Implemented an empty constructor for `BrowserHistoryBackup` class, removing the necessity to specify an initial backup path upon object creation. This change allows for more flexible instantiation of the backup utility class.

- Introduced the `SetupBackupPath` method for setting the backup path post-instantiation, providing a more intuitive setup flow and aligning with the enhanced user interaction model.

- Updated the `InvokeUserInteraction` method to include:
  - Prompting for a backup path if not already provided, ensuring a backup path is always set before proceeding.
  - Handling user input more gracefully, including validation checks for empty or whitespace-only paths.
  - Offering a selection of browsers for which the user wants to backup history, improving customization of the backup process.

- Fixed an issue in the selection process within `InvokeUserInteraction` where splitting the user's input and iterating over the selections did not properly handle each browser backup invocation.

- Corrected the file path concatenation in the `BackupChromeHistory` method to ensure compatibility and consistency across different environments.

These updates significantly improve the utility's flexibility, user-friendliness, and reliability, particularly in handling Firefox profiles and streamlining the user interaction process.

Date: March 10, 2024
Author: Tristan McGowan
Version: 0.1.1

+++ WinPreManage_browsers_history.ps1
@@ -68,16 +68,29 @@ class BrowserHistoryBackup {
     }
 
     [void]BackupFirefoxHistory() {
-        $firefoxProfilePath = Get-ChildItem "$env:APPDATA\Mozilla\Firefox\Profiles" -Directory | Select-Object -First 1
-        $firefoxHistoryPath = $firefoxProfilePath.FullName + "\places.sqlite"
-        if (Test-Path $firefoxHistoryPath) {
-            $destPath = Join-Path -Path $this.HistoryBackupPath -ChildPath "Firefox_History"
-            Copy-Item -Path $firefoxHistoryPath -Destination $destPath -Force
-            Write-Host "Firefox history backed up successfully to $destPath."
+        $firefoxProfilesPath = Join-Path -Path $env:APPDATA -ChildPath "Mozilla\Firefox\Profiles"
+        if (Test-Path $firefoxProfilesPath) {
+            $profileDirs = Get-ChildItem -Path $firefoxProfilesPath -Directory
+            $foundHistoryFile = $false
+
+            foreach ($dir in $profileDirs) {
+                $firefoxHistoryPath = Join-Path -Path $dir.FullName -ChildPath "places.sqlite"
+                if (Test-Path $firefoxHistoryPath) {
+                    $destPath = Join-Path -Path $this.HistoryBackupPath -ChildPath ($dir.Name + "_Firefox_History.sqlite")
+                    Copy-Item -Path $firefoxHistoryPath -Destination $destPath -Force
+                    Write-Host "Firefox history from profile '$($dir.Name)' backed up successfully to $destPath."
+                    $foundHistoryFile = $true
+                }
+            }
+
+            if (-not $foundHistoryFile) {
+                Write-Host "Firefox history database not found in any profile."
+            }
         } else {
-            Write-Host "Firefox history database not found."
+            Write-Host "Firefox profiles directory not found. Unable to locate Firefox profile directory."
         }
     }
+
 
     [void]BackupOperaHistory() {
         $operaHistoryPath = "$env:APPDATA\Opera Software\Opera Stable\History"
@@ -93,7 +106,7 @@ class BrowserHistoryBackup {
     [void]BackupChromeHistory() {
         $chromeHistoryPath = "$env:LOCALAPPDATA\Google\Chrome\User Data\Default\History"
         if (Test-Path $chromeHistoryPath) {
-            $destPath = Join-Path -Path $this.HistoryBackupPath -ChildPath "Chrome_History.txt"
+            $destPath = Join-Path -Path $this.HistoryBackupPath -ChildPath "Chrome_History"
             Copy-Item -Path $chromeHistoryPath -Destination $destPath -Force
             Write-Host "Chrome history backed up successfully to $destPath."
         } else {
